Added opportunity attacks. Added shove. Proably did a lot of other stuff

+++ Assets/Scripts/Scriptables/SupportAbilityData.cs
@@ -18,5 +18,5 @@ public class SupportAbilityData : Ability
     /// </summary>
     public int inDmgMod;
     public float hitMod;
-    public float dodgeMod;
+    public int dodgeMod;
 }
